Unit 3 lesson 3 uploaded

+++ Assignments/unit3-lesson2/uart.c
@@ -3,7 +3,7 @@
 #define UARTDR *((volatile unsigned int* const)((unsigned int*) 0x0101f1000))
 	void UART_TRANSMIT(unsigned  char* x){
 		while(*x !=0){
-			UARTDR = *x;
+			UARTDR = (unsigned int)(*x);
 			x++;
 		}
 	}
